+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.1] - 2026-03-25
+
+### Fixed
+- Fix init command error when commands directory doesn't exist
+- Remove commands copying logic (no longer needed)
+
+### Changed
+- Remove `--skip-commands` option from CLI
+
 ## [2.3.0] - 2026-03-25
 
 ### Added